hf_model.py
from huggingface_hub import InferenceClient
import os
from dotenv import load_dotenv

load_dotenv()

client = InferenceClient(
    model="mistralai/Mistral-7B-Instruct-v0.2",
    token=os.getenv("HF_API_TOKEN"),
)

def query_model(prompt):
    try:
        response = client.chat_completion(
            messages=[
                {"role": "user", "content": prompt}
            ],
            max_tokens=800,
        )

        return response.choices[0].message.content

    except Exception as e:
        return f"HuggingFace Error: {str(e)}"